Lecture 175 - Dependency Mapping
Important to understand the infrastructure of what needs what, and how they should be re-factored, or thought about prior to including. Don't have diamonds or loops, meaning one object shouldn't be able to talk to a different object through more than one different path. I.E. Controller talking to turret through Aiming or Tank. Only the Aiming should talk to the turret, and the Controller talking to the aiming.